Nkrafterz/requirements.txt
Django==5.0
djangorestframework==3.14.0
Jinja2==3.1.2